% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_mrp.R
\name{adjust_mrp}
\alias{adjust_mrp}
\title{Adjust an MRP model to match known aggregate outcomes}
\usage{
adjust_mrp(obj, nthin = 25)
}
\arguments{
\item{obj}{an object created by calling \code{hrr} with argument \code{mrp_only} set to TRUE}

\item{nthin}{number of iterations to keep}
}
\description{
\code{adjust_mrp} overwrites the random area intercepts contained
in an existing model.
}
