% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_support.R
\name{area_support}
\alias{area_support}
\title{Produce a tidy data frame with estimates of area support from a fit}
\usage{
area_support(object)
}
\arguments{
\item{object}{an object of class hrrfit}
}
\description{
Produce a tidy data frame with estimates of area support from a fit
}
