% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoprior.R
\name{autoprior}
\alias{autoprior}
\title{Generate brms priors given a formula and data}
\usage{
autoprior(formula, data)
}
\arguments{
\item{formula}{a formula object, or something that can be coerced to a formula object}

\item{data}{data frame containing the individual level data}
}
\value{
returns an object of class brmsprior
}
\description{
Generate brms priors given a formula and data
}
\examples{

}
