% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_dfs.R
\name{compare_dfs}
\alias{compare_dfs}
\title{Check two data frames contain specified variables with same class and levels}
\usage{
compare_dfs(formula, df1, df2)
}
\arguments{
\item{formula}{a one-sided formula specifying the variables to be checked}

\item{df1}{a data frame}

\item{df2}{a second data frame}
}
\value{
returns TRUE or reports an error
}
\description{
Check two data frames contain specified variables with same class and levels
}
\examples{
data(londonps)
data(londonmayor)
## This works
compare_dfs(~ageGroup, londonps, londonmayor)
## This complains about different classes
compare_dfs(~ageGroup + ethnicity, londonps, londonmayor)
londonmayor$ethnicity <- factor(londonmayor$ethnicity)
compare_dfs(~ageGroup + ethnicity, londonps, londonmayor)
}
