% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_support.R
\name{group_support}
\alias{group_support}
\title{Produce a tidy data frame with estimates of group support from a fit}
\usage{
group_support(object)
}
\arguments{
\item{object}{an object of class hrrfit}
}
\description{
Produce a tidy data frame with estimates of group support from a fit
}
