% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrr.R
\name{hrr}
\alias{hrr}
\title{Estimate hierarchical related regression model}
\usage{
hrr(formula, data, ps, result)
}
\arguments{
\item{formula}{a formula object, or something that can be coerced to a formula object}

\item{data}{data frame containing the individual level data}

\item{ps}{data frame containing the post-stratification data; must contain variable \code{count}}

\item{result}{data frame containing the results; must contain column names equal to levels of the dependent variable in \code{formula}}

\item{areavar}{a character containing the name of the variable giving the area}
}
\value{
returns TRUE or reports an error
}
\description{
Estimate hierarchical related regression model
}
