% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/londonmayoral.R
\docType{data}
\name{londonmayor}
\alias{londonmayor}
\title{Individual and aggregate data for the 2016 London Mayoral election}
\format{
A data frame with 1924 rows and 35 variables:
\describe{
\item{ONSCode}{Geographic identifier from the Office for National Statistics}
\item{Ward}{Name of the borough}
\item{ageGroup}{Respondent age group}
\item{education}{Respondent's highest level of educational qualifications}
\item{ethnicity}{Respondent ethnicity}
\item{gender}{Respondent gender}
\item{vi}{Respondent recalled vote, including non-voters}
\item{days_after_elex}{Fieldwork date minus date of election}
\item{ConPct, ConPct_mean, ConPct_sd, ConPct_sc}{Boris Johnson vote share in borough in 2012, together with global mean and SD, and scaled version}
\item{LabPct}{Ken Livingstone vote share in borough in 2012... }
\item{LDemPct}{Brian Paddick vote share in borough in 2012... }
\item{GreenPct}{Jenny Jones vote share in borough in 2012... }
\item{OtherPct}{Combined share of Siobhan Benita (Ind), Lawrence Webb (UKIP) and Carlos Cortiglia (BNP) vote share in borough in 2012... }
\item{Con_16_count}{Borough count of votes won by Zac Goldsmith}
\item{Lab_16_count}{Borough count of votes won by Sadiq Khan}
\item{Green_16_count}{Borough count of votes won by Sian Berry}
\item{LDem_16_count}{Borough count of votes won by Caroline Pidgeon}
\item{UKIP_16_count}{Borough count of votes won by Peter Whittle}
\item{Other_16_count}{Borough count of votes won by seven other candidates}
\item{DNV_16_count}{Borough population over 16 less total votes cast}
}
}
\source{
Individual data from the British Election Study. Results from 2012 and 2016 elections from \url{https://data.london.gov.uk/}
}
\usage{
londonmayor
}
\description{
A dataset containing individual post-election responses from the
British Election Study, auxiliary vote shares from 2012 mayoral
election, and aggregate counts from the 2016 election.
}
\keyword{datasets}
