% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{hrr}
\alias{hrr}
\alias{hrr-package}
\title{hrr provides a simple function, \code{\link{hr}}, to compose and print beautiful horizontal rules.}
\description{
hrr provides a simple function, \code{\link{hr}}, to
compose and print beautiful horizontal rules.
}
\section{Package options}{
  It uses the following \code{\link{options}} to configure
  behaviour: \itemize{ \item \code{hrr.symbol}: default
  character symbol }
}

