\name{ncols}
\alias{ncols}
\title{Computes the number of terminal window columns.}
\usage{
ncols(set_option = TRUE)
}
\arguments{
  \item{set_option}{whether to automatically set the R
  terminal width to the system terminal number of columns
  [default TRUE]}
}
\value{
Number of terminal window columns
}
\description{
Computes the number of terminal window columns.
}

